requirements.txt
# Web frameworks
flask==3.0.3
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
flask-sqlalchemy==3.1.1

# Auth
flask-login==0.6.3
werkzeug==3.0.3

# Email parsing
mail-parser==3.15.0
beautifulsoup4==4.12.3
lxml==5.2.2

# URL / domain intelligence
python-whois==0.9.4
dnspython==2.6.1
requests==2.32.3
urllib3==2.2.2

# Network scanning
python-nmap==0.7.1

# ML / AI
scikit-learn==1.5.0
transformers==4.41.2
torch==2.3.0
datasets==2.20.0
accelerate==0.31.0
huggingface-hub==0.23.4
joblib==1.4.2
numpy==1.26.4
scipy==1.13.1

# File analysis
yara-python==4.5.1
python-magic==0.4.27
olefile==0.47
pdfminer.six==20231228
python-docx==1.1.2

# Image analysis
opencv-python==4.10.0.84
pytesseract==0.3.10
Pillow==10.3.0

# NLP / text
langdetect==1.0.9

# Reporting / export
reportlab==4.2.2
fpdf2==2.7.9

# Utilities
python-dotenv==1.0.1
pydantic==2.7.4
tqdm==4.66.4
colorlog==6.8.2

# Testing
pytest==8.2.2
pytest-cov==5.0.0